Small standalone program to test wait() for pdfork()ed child
One of the purposes of the process descriptor functionality in Capsicum
is to allow libraries to use sandboxed sub-processes as part of their
implementation.  To avoid this being visible to the using application,
the process descriptor functionality therefore needs to avoid a situation
where "unexpected SIGCHLD signals are delivered to the application, and
unexpected process IDs are returned by wait" (quoting from the Capsicum
2010 Usenix paper).

This test program checks the latter aspect, by pdfork()ing a child and
when (normal) waitpid()ing for it afterwards.