Add simple SCTP test
On Linux, requires libsctp-dev package to build, and CONFIG_IP_SCTP
support in kernel configuration.

This commit also includes a makefile hack to allow different variable
contents between FreeBSD and Linux, while still keeping the master
makefile simple enough to run with either BSD make or GNU make.
 - With GNU make, the file GNUmakefile will be used by preference.  This
   detects Linux and sets Linux-specific variables (LIBSCTP at the moment),
   then includes the master makefile.
 - With BSD make, GNUmakefile is ignored and the normal makefile is read
   directly.

This works for now, but will not cope with a variable that needs a value on
FreeBSD but not on Linux...