casper: Don't use explicit group names / IDs in tests
The tests previously relied on the specific IDs/names of groups as
provided by FreeBSD, which is not portable.  Query for the first
few groups entries at start of day, and use those to filter rather than
explicit names/IDs.

For reference, the test previously relied on the following initial
groups entries from FreeBSD:
 0  wheel
 1 daemon
 2 kmem
 3 sys
 4 tty
 5 operator