Ensure that mknod(2) is called with S_IFIFO in `mode`
This prevents the test from failing with a false positive, i.e., -1/EINVAL,
as the file type was not specified prior to this commit. Other
alternatives (use S_IFBLK, S_IFCHR) is not possible, unless executing as
root, so use fifos instead (which can be done by all users.

This unbreaks the test on FreeBSD 13.0-CURRENT.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>