Create a temporary directory root at start if not specified via -T
One of the issues I noticed with the test suite is that it was
non-deterministic if interrupted, i.e., testcases that relied on a fresh
/tmp filesystem would fail on the second run if files were left behind.

Furthermore, some systems need `$TMPDIR` to not be hardcoded to `/tmp`,
e.g., ATF/Kyua (the defacto test runner/reporting infrastructure in FreeBSD
today). The fact that capsicum-test created files in `/tmp`, which was
outside of `$TMPDIR` [set by Kyua at runtime], means that it was
violating Kyua's sandboxing requirement.

This change uses tempnam to generate a temporary directory root at test
start via a `::testing::Environment` object (the global set of fixture hooks
provided by gtest), which it then uses to build other paths off of. If a
user provided path wasn't given, the directory root will be torn down at
test runner end, as best possible.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>