Fix `OpenatTest.WithFlag` when `O_BENEATH` is passed after 5eb909a37339fe4675ef95b769a07c5eb3894799
The absolute symlink failure to traverse testcases were buggy.

They were trying to verify that openat(2) would fail when opening a path
outside a sandbox, when in reality it was testing using a directory (`TOPDIR`)
fd that was in the sandbox. An easy to implement change was to instead test
the absolute path resolution failure using the subdirectory (`SUBDIR`)
fd.

Reindent the directory hierarchy comment and reorder the elements in the
comments and tests to support the change.

While here, rename `SUBDIR_ABS` to `SUBDIR`, as `SUBDIR` was not an
absolute path.

This fixes the rest of #28, along with
5eb909a37339fe4675ef95b769a07c5eb3894799.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>