Remove `FAIL` macro use for non-x86 architectures when testing `sysarch(2)`
`FAIL()` does not support being called in the form noted in the test,
which causes a test failure on non-x86 architectures.

The alternatives (use `ADD_TEST_FAILURE()` or `GTEST_SKIP()`) would be
misleading (in both cases), and in the case of `GTEST_SKIP()` is unavailable
on the version of googletest packaged with capsicum-test.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>