Allow `capsicum-test` to be executed from an absolute path
The proposed change stashes the execution directory for `capsicum-test`,
then uses it to compute the directory where the `mini-me*` binaries are
expected to live.

This change required refactoring how the fexecve tests were executed,
as the path for the `mini-me*` binaries could no longer be determined at
compile-time, but needed to be determined at runtime. As such, the code was
switched from `FORK_TEST(..)` to `FORK_TEST_F(..)` -- with an appropriate
class setting the path to `mini-me*` in the constructor, and
`FORK_TEST_ON(..)` to `FORK_F(..)`, building off the class developed for
`FORK_TEST_F(..)` with proper setup/teardown fixtures for cleaning up the
temporary script.

As an added bonus, the temporary script is guaranteed to not be around
anymore at test execution completion.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>