Check capsicum feature prerequisites on FreeBSD before running tests
The capsicum-test test suite assumes that it's always being run on a
CAPABILITIES enabled FreeBSD kernel. However, there's a chance that the test
will be run on a host without a CAPABILITIES enabled kernel, so we must check
for the support before running the tests. The way to achieve this is to add
a `feature_present("security_capabilities")` check to
`SetupEnvironment::SetUp()`, and skip the tests when the support is not
available.

While here, add a check for `kern.trap_enotcap` being enabled in the kernel.
As noted by @markjdb in #23, the sysctl being enabled can trigger
non-deterministic failures. Therefore, the tests should be skipped if this
sysctl is enabled.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>