Shift to use GoogleTest's skipped test tracking
Drop local hand-rolled mechanism for skipping tests, and instead
use GTEST_SKIP().

Rename test cases that have a condition to include "..If<Condition>"
in their names, so it's clear from the final output why a test might
be skipped.

Split the Capability.SyscallAt test into the parts that need root
and the parts that don't.