Improve failure mode if other_uid cannot be inferred (#45)
If the capsicum-test binary is owned by root, we would infer other_uid as
root. This caused the tests that require another non-root UID to fail
because root can bypass some of the checks that would otherwise be in place.

To avoid surprising test failures, add an ASSERT_NE(0u, other_uid) with a
suggestion on how to fix this problem.