Use libprocstat to get child process state on FreeBSD
I have been seeing intermittent failures in the CheriBSD that seem to
be happening due to the `signal(SIGCHLD, SIG_IGN)` call inside
ProcessState() that leads to child process not entering zombie state.

To avoid this race use libprocstat instead of forking a new ps process.
This is not only more reliable, but should also speed up the tests a bit
when running on slow emulated platforms.