Use volatile sig_atomic_t for signal handler variables
I was trying to update the Pdkill test to no longer use sleep() and noticed
that the read of the had_signal map was being optimized out by the compiler.
Writing to a std::map inside a signal handler is not safe, so replace it
with a sig_atomic_t array.