Verify some more socket operations
Currently we allow creation of raw sockets and routing sockets in
capability mode.  This violates some of Capsicum's guarantees.

On FreeBSD, socket ioctls are also the primary interface for configuring
network interfaces, and they are not restricted in capability mode.
Because we lack a sane capability-based system call interface for
manipulating the namespace of network interfaces, all such ioctls should
simply be prohibited.

Add some simple regression tests for these cases in preparation for
fixing the problem in FreeBSD.